rpi_zero_agent/requirements.txt
pyserial>=3.5
paho-mqtt>=1.6.0
requests>=2.25.0
